AE-2006: Add hallinto-oikeus selection column to notice-bailiff-osapuolet-table.svelte
- Can be shown or hidden with a boolean prop
- Delete penalty-notice-bailiff-osapuolet-table.svelte
