+++ src/language/sv.json
@@ -1871,7 +1871,7 @@
         "cancel-button": "Återkalla",
         "preview-button": "Förhandsvisa",
         "download-button": "Ladda ner",
-        "court": "Hallinto-oikeuden valinta (sv)",
+        "court": "Hallinto-oikeus (sv)",
         "manually-sent": "Valitun toimenpiteen asiakirja tulee toimittaa manuaalisesti (sv)",
         "no-delivery": "Dokumenttia ei toimiteta osapuolelle (sv)",
         "case": {
@@ -2012,7 +2012,6 @@
         "penalty-decision-notice-bailiff": {
           "title": "Tiedoksianto haastemiehelle (sv)",
           "info": "Sakkopäätös toimitetaan kirjaamon kautta haastemiehelle, jos toinen postitus palautuu takaisin. (sv)",
-          "court": "Hallinto-oikeus (sv)",
           "description": "Kommentti (sv)",
           "publish-button": "Luo toimenpide (sv)",
           "district-court": "Käräjäoikeus (sv)",
